+++ R/jamenrich-plot.R
@@ -153,7 +153,7 @@ mem_gene_path_heatmap <- function
    ## pathways with min_set_ct_each also meet p_cutoff
    met_p_cutoff <- (mem$enrichIM[colnames(memIM),,drop=FALSE] <= p_cutoff)
    met_min_set_ct_each <- do.call(cbind, lapply(nameVector(colnames(mem$geneIM)), function(icol){
-      colSums(mem$geneIM[rownames(memIM),icol] * memIM) >= min_set_ct_each;
+      colSums(mem$geneIM[rownames(memIM),icol] * (memIM != 0)) >= min_set_ct_each;
    }));
    met_criteria <- rowSums(met_p_cutoff & met_min_set_ct_each) > 0;
    if (any(!met_criteria) && verbose) {